+++ NEWS.md
@@ -1,3 +1,8 @@
+# ICAMS 2.2.5.90xx
+## Changed
+* Updated function `MakeDataFrameFromVCF` to use `data.table::fread` instead of 
+`read.csv`.
+
 # ICAMS 2.2.5 
 ## Changed
 * Updated function `CreateOneColSBSMatrix` to throw a message instead of an error when there are SBS variant whose reference base in ref.genome does not match the reference base in the VCF file.